+++ NEWS.md
@@ -1,12 +1,36 @@
 # yardstick development
 
+## New metrics and functionality
+
+* `average_precision()` is a probability metric that can be used as an alternative to `pr_auc()`. It has the benefit of avoiding any issues of ambiguity in the case where `recall == 0` and the current number of false positives is `0`.
+
 ## Other improvements
+
+* The `autoplot()` method for `pr_curve()` has been improved to always set the
+axis limits to `c(0, 1)`.
 
 * All valid arguments to `pROC::roc()` are now utilized, including those passed
 on to `pROC::auc()`.
 
 * Documentation for class probability metrics has been improved with
 more informative examples (@rudeboybert, #100).
+
+## Bug fixes
+
+* `pr_curve()` now places a `1` as the first precision value, rather than a
+`NA`. While `NA` is technically correct as precision is undefined here, `1` is
+practically more correct because it generates a correct PR Curve graph and, 
+more importantly, allows `pr_auc()` to compute the correct AUC.
+
+* `pr_curve()` could generate the wrong results in the somewhat rare case 
+when two class probability estimates were the same, but had different 
+truth values.
+
+* `pr_curve()` (and subsequently `pr_auc()`) now generates the correct curve
+when there are duplicate class probability values (reported by @dariyasydykova, #93).
+
+* Binary `mcc()` now avoids integer overflow when the confusion matrix 
+elements are large (#108).
 
 # yardstick 0.0.3